Consolidate BroadcastManager and MainComponent.
